BlockChain/requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dnspython==1.16.0
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pymongo==3.11.0
pypng==0.0.20
PyQRCode==1.2.1
requests==2.24.0
urllib3==1.25.10
Werkzeug==1.0.1
gunicorn==20.0.4
